requirements.txt
Flask==2.3.2
PyPDF2==2.12.1
python-docx==0.8.11
scikit-learn==1.1.2
Werkzeug==2.3.2
